SC-4331: only reset codec if driver module is not loaded
Resetting the codec when the driver is loaded results in the registers
being reset back to defaults but the driver view is still of the
patched versions of the registers.

Rerunning audio.sh is only done when debugging so this will be less
confusing for the debug case and never an issue for the normal case.

Also added some debug code to dump regs before and after the fixup.
Registers land in /tmp/cxregs.  We can remove this later after we are
certain all is working fine.