SC-6915: add -E nodiscard to mkfs calls
This avoids the emmc getting a TRIM call for all the empty blocks which
take forever to erase.  This is fine for new emmc because all blocks are
alread erased.

Note: For reflashing this can make future write speeds slower
if the emmc has to erase blocks on the fly.